Amazon/requirements.txt
Flask==2.3.3
Flask-MySQL==1.5.2
Flask-Session==0.5.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
Pillow==10.0.1
razorpay==1.3.0
python-dotenv==1.0.0
PyMySQL==1.1.0
bcrypt==4.0.1
openai==1.3.0
requests==2.31.0